+++ packages/@guardian/src-button/LinkButton.stories.tsx
@@ -1,7 +1,7 @@
 import { LinkButton, LinkButtonProps } from './LinkButton';
 import { SvgArrowRightStraight } from '@guardian/src-icons';
-import { asPlayground, asChromaticStory } from '../../../../lib/story-intents';
-import type { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
+import { asPlayground, asChromaticStory } from '../../../lib/story-intents';
+import type { Story } from '../../../lib/@types/storybook-emotion-10-fixes';
 // These types are the right types, but don't work with Storybook v6 which uses Emotion v10
 // import type { Args, Story } from '@storybook/react';
