+++ packages/@guardian/src-button/tsconfig.json
@@ -1,18 +1,18 @@
 {
-  "extends": "../../../../tsconfig.json",
+  "extends": "../../../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist/types",
     "declaration": true,
     "emitDeclarationOnly": true
   },
-  "include": [".", "../../foundations/types/themes.d.ts"],
+  "include": [".", "../../../src/core/foundations/types/themes.d.ts"],
   "exclude": ["dist", "*.stories.tsx"],
   "references": [
     {
-      "path": "../../foundations"
+      "path": "../../../src/core/foundations"
     },
     {
-      "path": "../../icons"
+      "path": "../../../src/core/icons"
     }
   ]
 }
